requirements.txt
asgiref==3.8.1
chardet==5.2.0
Django==5.2
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
et_xmlfile==2.0.0
flake8==7.2.0
mccabe==0.7.0
modeltranslation==0.25
openpyxl==3.1.5
pillow==11.2.1
psycopg2-binary==2.9.10
pycodestyle==2.13.0
pyflakes==3.3.2
python-dotenv==1.1.0
reportlab==4.4.0
sqlparse==0.5.3
tzdata==2025.2
XlsxWriter==3.2.3
python-decouple==3.8
gunicorn==23.0.0